+++ sos-webservices/sos-webservices-joc/src/main/java/com/sos/joc/publish/util/ExportUtils.java
@@ -320,7 +320,7 @@ public class ExportUtils {
             configurationObject.setPath(item.getPath());
             configurationObject.setName(item.getName());
             configurationObject.setObjectType(ConfigurationType.fromValue(item.getType()));
-            configurationObject.setConfiguration(JocInventory.content2IJSObject(item.getInvContent(), configurationObject.getObjectType().intValue()));
+            configurationObject.setConfiguration(JocInventory.content2IJSObject(item.getInvContent(), configurationObject.getObjectType()));
             return configurationObject;
         } catch (IOException e) {
             throw new JocException(e);
