+++ sos-webservices/sos-webservices-joc/src/main/java/com/sos/joc/workflow/impl/WorkflowOrderTemplatesImpl.java
@@ -57,7 +57,7 @@ public class WorkflowOrderTemplatesImpl extends JOCResourceImpl implements IWork
             
             entity.setSchedules(schedules.stream().filter(hasOrderParameterisation).map(item -> {
                 try {
-                    Schedule s = (Schedule) JocInventory.content2IJSObject(item.getContent(), ConfigurationType.SCHEDULE.intValue());
+                    Schedule s = (Schedule) JocInventory.content2IJSObject(item.getContent(), ConfigurationType.SCHEDULE);
                     if (s.getOrderParameterisations() != null && !s.getOrderParameterisations().isEmpty()) {
                         if ((s.getWorkflowName() != null && workflame.equals(s.getWorkflowName())) || (s.getWorkflowNames() != null && s
                                 .getWorkflowNames().contains(workflame))) {
